+++ extension/setSelection.js
@@ -1,0 +1,12 @@
+module.exports = async (context, { newSelection }) => {
+  try {
+    await context.storage.device.set('selection', newSelection)
+      .then(
+        context.log.info(`New selection is '${newSelection.path}'`)
+      );
+  } catch (err) {
+    context.log.error(err, 'Failed to set new selection');
+
+    throw err;
+  }
+};