+++ frontend/action-creators/index.js
@@ -1,22 +1,44 @@
 import {
-  SET_SWITCH_SELECTION,
-  ERROR_SWITCH_SELECTION,
+  REQUEST_SELECTION,
+  RECEIVE_SELECTION,
+  SET_SELECTION,
+  ERROR_SELECTION,
 } from '../constants';
 
 /**
+ * Start selection request
+ * @return {Object}
+ */
+export const requestSelection = () => ({
+  type: REQUEST_SELECTION,
+});
+
+/**
+ * Recieve selection
  * @param {Object} selection .
  * @return {Object}
  */
-export const setSwitchSelection = selection => ({
-  type: SET_SWITCH_SELECTION,
+export const receiveSelection = selection => ({
+  type: RECEIVE_SELECTION,
   selection,
 });
 
 /**
- * @param {Error} error .
+ * Set selection
+ * @param {Object} selection .
  * @return {Object}
  */
-export const errorSwitchSelection = error => ({
-  type: ERROR_SWITCH_SELECTION,
+export const setSelection = selection => ({
+  type: SET_SELECTION,
+  selection,
+});
+
+/**
+ * Error selection
+ * @param {Object} error .
+ * @return {Object}
+ */
+export const errorSelection = error => ({
+  type: ERROR_SELECTION,
   error,
 });