+++ frontend/components/SwitchHeader/SwitchButton/connector.js
@@ -1,6 +1,5 @@
 import { connect } from 'react-redux';
-import { getSelection } from '../../../selectors';
-import { makeUpdateSelection } from '../../../actions';
+import { updateSelection } from '../../../actions';
 
 /**
  * Maps the contents of the state to the component props.
@@ -8,16 +7,7 @@ import { makeUpdateSelection } from '../../../actions';
  * @return {Object}
  */
 const mapDispatchToProps = dispatch => ({
-  setSelection: path => dispatch(makeUpdateSelection(path)),
+  setSelection: link => dispatch(updateSelection(link)),
 });
 
-/**
- * Maps the contents of the state to the component props.
- * @param {Object} state The current application state.
- * @return {Object} The extended component props.
- */
-const makeMapStateToProps = state => ({
-  isSelected: getSelection(state),
-});
-
-export default connect(makeMapStateToProps, mapDispatchToProps);
+export default connect(null, mapDispatchToProps);