+++ frontend/components/SwitchHeader/SwitchButton/index.jsx
@@ -1,40 +1,35 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
 import connect from './connector';
 import styles from './style';
 
 /**
  * SwitchButton component
- * @param {Object} link The first number.
+ * @param {Object} props The component props.
+ * @param {JSX} .
  */
-class SwitchButton extends Component {
-  static propTypes = {
-    isActive: PropTypes.bool.isRequired,
-    link: PropTypes.shape().isRequired,
-    setSelection: PropTypes.func.isRequired,
+const SwitchButton = ({ isActive, link, setSelection }) => {
+  const handleClick = () => {
+    setSelection(link);
   };
 
-  static defaultProps = {};
+  return (
+    <button
+      type="button"
+      onClick={handleClick}
+      className={isActive ? styles.activeButton : styles.button}
+    >
+      {link.label}
+    </button>
+  );
+};
 
-  handleClick = (link) => {
-    this.props.setSelection(link);
-  }
+SwitchButton.propTypes = {
+  isActive: PropTypes.bool.isRequired,
+  link: PropTypes.shape().isRequired,
+  setSelection: PropTypes.func.isRequired,
+};
 
-  /**
-   * Render menu item button
-   * @returns {JSX}
-   */
-  render() {
-    return (
-      <button
-        type="button"
-        onClick={() => this.handleClick(this.props.link)}
-        className={this.props.isActive ? styles.activeButton : styles.button}
-      >
-        {this.props.link.label}
-      </button>
-    );
-  }
-}
+SwitchButton.defaultProps = {};
 
 export default connect(SwitchButton);