+++ CHANGELOG.md
@@ -5,8 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [1.0.0] - 2023-11-10
 ### Added
 - Adds a page switcher to the header
-- The selected page will be the 'home page'
-
+- The selected page will be the start page on the next start
 
 [Unreleased]: https://github.com/shopgate/ext-cliplister/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/shopgate/ext-cliplister/compare/v0.0.1...v1.0.0